Fix issue with adding dependent entities
