+++ src/backend/Csrs.Interfaces.Dynamics/ModelExtensions/ChildIsDependent.cs
@@ -1,0 +1,12 @@
+﻿namespace Csrs.Interfaces.Dynamics.Models
+{
+    /// <summary>
+    /// Picklist values for ssg_csrschild.ssg_childisadependent
+    /// </summary>
+    public enum ChildIsDependent
+    {
+        Yes = 867670000,
+        No = 867670001,
+        DontKnow = 867670002
+    }
+}
