+++ src/backend/Csrs.Interfaces.Dynamics/ModelExtensions/MicrosoftDynamicsCRMssgCsrschild.cs
@@ -1,0 +1,23 @@
+﻿using Newtonsoft.Json;
+using System.ComponentModel.DataAnnotations;
+
+namespace Csrs.Interfaces.Dynamics.Models
+{
+    public partial class MicrosoftDynamicsCRMssgCsrschildMetadata
+    {
+        [JsonProperty(PropertyName = "ssg_dateofbirth")]
+        [JsonConverter(typeof(DateFormatConverter), "yyyy-MM-dd")]
+        public DateTimeOffset? SsgDateofbirth { get; set; }
+    }
+
+    [MetadataType(typeof(MicrosoftDynamicsCRMssgCsrschildMetadata))]
+    public partial class MicrosoftDynamicsCRMssgCsrschild
+    {
+        [JsonProperty(PropertyName = "ssg_ChildsFather@odata.bind")]
+        public string SsgChildsFatherODataBind { get; set; }
+
+        [JsonProperty(PropertyName = "ssg_ChildsMother@odata.bind")]
+        public string ssg_ChildsMotherODataBind { get; set; }
+
+    }
+}
