+++ src/backend/Csrs.Interfaces.Dynamics/Models/MicrosoftDynamicsCRMssgCsrschild.cs
@@ -10,8 +10,6 @@ namespace Csrs.Interfaces.Dynamics.Models
     using System.Collections;
     using System.Collections.Generic;
     using System.Linq;
-    using System.Dynamic;
-    
 
     /// <summary>
     /// Microsoft.Dynamics.CRM.ssg_csrschild
@@ -31,8 +29,7 @@ namespace Csrs.Interfaces.Dynamics.Models
         /// Initializes a new instance of the MicrosoftDynamicsCRMssgCsrschild
         /// class.
         /// </summary>
-        //public MicrosoftDynamicsCRMssgCsrschild(string _owninguserValue = default(string), string _ssgChildsfatherValue = default(string), string _owningteamValue = default(string), int? statuscode = default(int?), System.DateTimeOffset? overriddencreatedon = default(System.DateTimeOffset?), string ssgMiddlename = default(string), string ssgLastname = default(string), string _ssgFilenumberValue = default(string), string _owneridValue = default(string), string _modifiedonbehalfbyValue = default(string), string _createdonbehalfbyValue = default(string), string ssgCsrschildid = default(string), int? ssgChildisadependent = default(int?), string _owningbusinessunitValue = default(string), bool? ssgChildatageofmajority = default(bool?), string ssgTriggerrecalcrelationship = default(string), string _createdbyValue = default(string), System.DateTimeOffset? modifiedon = default(System.DateTimeOffset?), string _ssgChildsmotherValue = default(string), System.DateTimeOffset? ssgDateofbirth = default(System.DateTimeOffset?), int? utcconversiontimezonecode = default(int?), int? statecode = default(int?), string ssgFirstname = default(string), string ssgFullname = default(string), int? importsequencenumber = default(int?), int? timezoneruleversionnumber = default(int?), string versionnumber = default(string), System.DateTimeOffset? createdon = default(System.DateTimeOffset?), string _modifiedbyValue = default(string), MicrosoftDynamicsCRMsystemuser createdby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser createdonbehalfby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser modifiedby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser modifiedonbehalfby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser owninguser = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMteam owningteam = default(MicrosoftDynamicsCRMteam), MicrosoftDynamicsCRMprincipal ownerid = default(MicrosoftDynamicsCRMprincipal), MicrosoftDynamicsCRMbusinessunit owningbusinessunit = default(MicrosoftDynamicsCRMbusinessunit), IList<MicrosoftDynamicsCRMsyncerror> ssgCsrschildSyncErrors = default(IList<MicrosoftDynamicsCRMsyncerror>), IList<MicrosoftDynamicsCRMduplicaterecord> ssgCsrschildDuplicateMatchingRecord = default(IList<MicrosoftDynamicsCRMduplicaterecord>), IList<MicrosoftDynamicsCRMduplicaterecord> ssgCsrschildDuplicateBaseRecord = default(IList<MicrosoftDynamicsCRMduplicaterecord>), IList<MicrosoftDynamicsCRMasyncoperation> ssgCsrschildAsyncOperations = default(IList<MicrosoftDynamicsCRMasyncoperation>), IList<MicrosoftDynamicsCRMmailboxtrackingfolder> ssgCsrschildMailboxTrackingFolders = default(IList<MicrosoftDynamicsCRMmailboxtrackingfolder>), IList<MicrosoftDynamicsCRMprocesssession> ssgCsrschildProcessSession = default(IList<MicrosoftDynamicsCRMprocesssession>), IList<MicrosoftDynamicsCRMbulkdeletefailure> ssgCsrschildBulkDeleteFailures = default(IList<MicrosoftDynamicsCRMbulkdeletefailure>), IList<MicrosoftDynamicsCRMprincipalobjectattributeaccess> ssgCsrschildPrincipalObjectAttributeAccesses = default(IList<MicrosoftDynamicsCRMprincipalobjectattributeaccess>), MicrosoftDynamicsCRMssgCsrsparty ssgChildsMother = default(MicrosoftDynamicsCRMssgCsrsparty), MicrosoftDynamicsCRMssgCsrsparty ssgChildsFather = default(MicrosoftDynamicsCRMssgCsrsparty), MicrosoftDynamicsCRMssgCsrsfile ssgFileNumber = default(MicrosoftDynamicsCRMssgCsrsfile), IList<MicrosoftDynamicsCRMsharepointdocumentlocation> ssgCsrschildSharePointDocumentLocations = default(IList<MicrosoftDynamicsCRMsharepointdocumentlocation>), IList<MicrosoftDynamicsCRMssgCsrsrecalculation> ssgSsgCsrschildSsgCsrsrecalculation = default(IList<MicrosoftDynamicsCRMssgCsrsrecalculation>))
-        public MicrosoftDynamicsCRMssgCsrschild(string _owninguserValue = default(string), string _ssgChildsfatherValue = default(string), string _owningteamValue = default(string), int? statuscode = default(int?), System.DateTimeOffset? overriddencreatedon = default(System.DateTimeOffset?), string ssgMiddlename = default(string), string ssgLastname = default(string), string _ssgFilenumberValue = default(string), string _owneridValue = default(string), string _modifiedonbehalfbyValue = default(string), string _createdonbehalfbyValue = default(string), string ssgCsrschildid = default(string), int? ssgChildisadependent = default(int?), string _owningbusinessunitValue = default(string), bool? ssgChildatageofmajority = default(bool?), string ssgTriggerrecalcrelationship = default(string), string _createdbyValue = default(string), System.DateTimeOffset? modifiedon = default(System.DateTimeOffset?), string _ssgChildsmotherValue = default(string), string ssgDateofbirth = default(string), int? utcconversiontimezonecode = default(int?), int? statecode = default(int?), string ssgFirstname = default(string), string ssgFullname = default(string), int? importsequencenumber = default(int?), int? timezoneruleversionnumber = default(int?), int? versionnumber = default(int?), System.DateTimeOffset? createdon = default(System.DateTimeOffset?), string _modifiedbyValue = default(string), MicrosoftDynamicsCRMsystemuser createdby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser createdonbehalfby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser modifiedby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser modifiedonbehalfby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser owninguser = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMteam owningteam = default(MicrosoftDynamicsCRMteam), MicrosoftDynamicsCRMprincipal ownerid = default(MicrosoftDynamicsCRMprincipal), MicrosoftDynamicsCRMbusinessunit owningbusinessunit = default(MicrosoftDynamicsCRMbusinessunit), IList<MicrosoftDynamicsCRMsyncerror> ssgCsrschildSyncErrors = default(IList<MicrosoftDynamicsCRMsyncerror>), IList<MicrosoftDynamicsCRMduplicaterecord> ssgCsrschildDuplicateMatchingRecord = default(IList<MicrosoftDynamicsCRMduplicaterecord>), IList<MicrosoftDynamicsCRMduplicaterecord> ssgCsrschildDuplicateBaseRecord = default(IList<MicrosoftDynamicsCRMduplicaterecord>), IList<MicrosoftDynamicsCRMasyncoperation> ssgCsrschildAsyncOperations = default(IList<MicrosoftDynamicsCRMasyncoperation>), IList<MicrosoftDynamicsCRMmailboxtrackingfolder> ssgCsrschildMailboxTrackingFolders = default(IList<MicrosoftDynamicsCRMmailboxtrackingfolder>), IList<MicrosoftDynamicsCRMprocesssession> ssgCsrschildProcessSession = default(IList<MicrosoftDynamicsCRMprocesssession>), IList<MicrosoftDynamicsCRMbulkdeletefailure> ssgCsrschildBulkDeleteFailures = default(IList<MicrosoftDynamicsCRMbulkdeletefailure>), IList<MicrosoftDynamicsCRMprincipalobjectattributeaccess> ssgCsrschildPrincipalObjectAttributeAccesses = default(IList<MicrosoftDynamicsCRMprincipalobjectattributeaccess>), MicrosoftDynamicsCRMssgCsrsparty ssgChildsMother = default(MicrosoftDynamicsCRMssgCsrsparty), MicrosoftDynamicsCRMssgCsrsparty ssgChildsFather = default(MicrosoftDynamicsCRMssgCsrsparty), MicrosoftDynamicsCRMssgCsrsfile ssgFileNumber = default(MicrosoftDynamicsCRMssgCsrsfile), IList<MicrosoftDynamicsCRMsharepointdocumentlocation> ssgCsrschildSharePointDocumentLocations = default(IList<MicrosoftDynamicsCRMsharepointdocumentlocation>), IList<MicrosoftDynamicsCRMssgCsrsrecalculation> ssgSsgCsrschildSsgCsrsrecalculation = default(IList<MicrosoftDynamicsCRMssgCsrsrecalculation>))
+        public MicrosoftDynamicsCRMssgCsrschild(string _owninguserValue = default(string), string _ssgChildsfatherValue = default(string), string _owningteamValue = default(string), int? statuscode = default(int?), System.DateTimeOffset? overriddencreatedon = default(System.DateTimeOffset?), string ssgMiddlename = default(string), string ssgLastname = default(string), string _ssgFilenumberValue = default(string), string _owneridValue = default(string), string _modifiedonbehalfbyValue = default(string), string _createdonbehalfbyValue = default(string), string ssgCsrschildid = default(string), int? ssgChildisadependent = default(int?), string _owningbusinessunitValue = default(string), bool? ssgChildatageofmajority = default(bool?), string ssgTriggerrecalcrelationship = default(string), string _createdbyValue = default(string), System.DateTimeOffset? modifiedon = default(System.DateTimeOffset?), string _ssgChildsmotherValue = default(string), System.DateTimeOffset? ssgDateofbirth = default(System.DateTimeOffset?), int? utcconversiontimezonecode = default(int?), int? statecode = default(int?), string ssgFirstname = default(string), string ssgFullname = default(string), int? importsequencenumber = default(int?), int? timezoneruleversionnumber = default(int?), string versionnumber = default(string), System.DateTimeOffset? createdon = default(System.DateTimeOffset?), string _modifiedbyValue = default(string), MicrosoftDynamicsCRMsystemuser createdby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser createdonbehalfby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser modifiedby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser modifiedonbehalfby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser owninguser = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMteam owningteam = default(MicrosoftDynamicsCRMteam), MicrosoftDynamicsCRMprincipal ownerid = default(MicrosoftDynamicsCRMprincipal), MicrosoftDynamicsCRMbusinessunit owningbusinessunit = default(MicrosoftDynamicsCRMbusinessunit), IList<MicrosoftDynamicsCRMsyncerror> ssgCsrschildSyncErrors = default(IList<MicrosoftDynamicsCRMsyncerror>), IList<MicrosoftDynamicsCRMduplicaterecord> ssgCsrschildDuplicateMatchingRecord = default(IList<MicrosoftDynamicsCRMduplicaterecord>), IList<MicrosoftDynamicsCRMduplicaterecord> ssgCsrschildDuplicateBaseRecord = default(IList<MicrosoftDynamicsCRMduplicaterecord>), IList<MicrosoftDynamicsCRMasyncoperation> ssgCsrschildAsyncOperations = default(IList<MicrosoftDynamicsCRMasyncoperation>), IList<MicrosoftDynamicsCRMmailboxtrackingfolder> ssgCsrschildMailboxTrackingFolders = default(IList<MicrosoftDynamicsCRMmailboxtrackingfolder>), IList<MicrosoftDynamicsCRMprocesssession> ssgCsrschildProcessSession = default(IList<MicrosoftDynamicsCRMprocesssession>), IList<MicrosoftDynamicsCRMbulkdeletefailure> ssgCsrschildBulkDeleteFailures = default(IList<MicrosoftDynamicsCRMbulkdeletefailure>), IList<MicrosoftDynamicsCRMprincipalobjectattributeaccess> ssgCsrschildPrincipalObjectAttributeAccesses = default(IList<MicrosoftDynamicsCRMprincipalobjectattributeaccess>), MicrosoftDynamicsCRMssgCsrsparty ssgChildsMother = default(MicrosoftDynamicsCRMssgCsrsparty), MicrosoftDynamicsCRMssgCsrsparty ssgChildsFather = default(MicrosoftDynamicsCRMssgCsrsparty), MicrosoftDynamicsCRMssgCsrsfile ssgFileNumber = default(MicrosoftDynamicsCRMssgCsrsfile), IList<MicrosoftDynamicsCRMsharepointdocumentlocation> ssgCsrschildSharePointDocumentLocations = default(IList<MicrosoftDynamicsCRMsharepointdocumentlocation>), IList<MicrosoftDynamicsCRMssgCsrsrecalculation> ssgSsgCsrschildSsgCsrsrecalculation = default(IList<MicrosoftDynamicsCRMssgCsrsrecalculation>))
         {
             this._owninguserValue = _owninguserValue;
             this._ssgChildsfatherValue = _ssgChildsfatherValue;
@@ -190,8 +187,7 @@ namespace Csrs.Interfaces.Dynamics.Models
         /// <summary>
         /// </summary>
         [JsonProperty(PropertyName = "ssg_dateofbirth")]
-        //public System.DateTimeOffset? SsgDateofbirth { get; set; }
-        public string? SsgDateofbirth { get; set; }
+        public System.DateTimeOffset? SsgDateofbirth { get; set; }
 
         /// <summary>
         /// </summary>
@@ -226,7 +222,7 @@ namespace Csrs.Interfaces.Dynamics.Models
         /// <summary>
         /// </summary>
         [JsonProperty(PropertyName = "versionnumber")]
-        public int? Versionnumber { get; set; }
+        public string Versionnumber { get; set; }
 
         /// <summary>
         /// </summary>
