+++ src/backend/Csrs.Interfaces.Dynamics/Models/MicrosoftDynamicsCRMssgCsrsparty.cs
@@ -29,7 +29,7 @@ namespace Csrs.Interfaces.Dynamics.Models
         /// Initializes a new instance of the MicrosoftDynamicsCRMssgCsrsparty
         /// class.
         /// </summary>
-        public MicrosoftDynamicsCRMssgCsrsparty(int? statecode = default(int?), int? ssgStagingfilenumber = default(int?), string _ssgDuplicatedpartyValue = default(string), System.DateTimeOffset? createdon = default(System.DateTimeOffset?), int? ssgIncomeassistance = default(int?), int? ssgReferral = default(int?), string ssgCellphone = default(string), string ssgFirstname = default(string), int? timezoneruleversionnumber = default(int?), System.DateTimeOffset? modifiedon = default(System.DateTimeOffset?), int? importsequencenumber = default(int?), string ssgCsrspartyid = default(string), string _createdonbehalfbyValue = default(string), int? ssgPartyalreadyexists = default(int?), string _modifiedbyValue = default(string), int? utcconversiontimezonecode = default(int?), string? ssgDateofbirth = default(string), System.DateTimeOffset? overriddencreatedon = default(System.DateTimeOffset?), int? versionnumber = default(int?), string ssgLastname = default(string), string ssgReferencenumber = default(string), int? ssgPartygender = default(int?), string ssgBceidUserid = default(string), int? ssgPreferredcontactmethod = default(int?), string ssgCity = default(string), string ssgWorkphone = default(string), int? ssgProvinceterritory = default(int?), string _owningbusinessunitValue = default(string), string _owningteamValue = default(string), string ssgMiddlename = default(string), string ssgFullname = default(string), int? ssgGender = default(int?), string ssgHomephone = default(string), string _owninguserValue = default(string), int? statuscode = default(int?), string _modifiedonbehalfbyValue = default(string), string _ssgLatestduplicatepartyValue = default(string), string ssgStreet2 = default(string), string _createdbyValue = default(string), int? ssgStagingid = default(int?), string _owneridValue = default(string), string ssgAreapostalcode = default(string), string ssgEmail = default(string), System.DateTimeOffset? ssgBceidLastUpdate = default(System.DateTimeOffset?), string ssgPreferredname = default(string), bool? ssgPortalaccess = default(bool?), int? ssgIdentity = default(int?), string ssgBceidDisplayname = default(string), string ssgIdentityotherdetails = default(string), string ssgStreet1 = default(string), string ssgBceidGuid = default(string), MicrosoftDynamicsCRMsystemuser createdby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser createdonbehalfby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser modifiedby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser modifiedonbehalfby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser owninguser = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMteam owningteam = default(MicrosoftDynamicsCRMteam), MicrosoftDynamicsCRMprincipal ownerid = default(MicrosoftDynamicsCRMprincipal), MicrosoftDynamicsCRMbusinessunit owningbusinessunit = default(MicrosoftDynamicsCRMbusinessunit), IList<MicrosoftDynamicsCRMsyncerror> ssgCsrspartySyncErrors = default(IList<MicrosoftDynamicsCRMsyncerror>), IList<MicrosoftDynamicsCRMduplicaterecord> ssgCsrspartyDuplicateMatchingRecord = default(IList<MicrosoftDynamicsCRMduplicaterecord>), IList<MicrosoftDynamicsCRMduplicaterecord> ssgCsrspartyDuplicateBaseRecord = default(IList<MicrosoftDynamicsCRMduplicaterecord>), IList<MicrosoftDynamicsCRMasyncoperation> ssgCsrspartyAsyncOperations = default(IList<MicrosoftDynamicsCRMasyncoperation>), IList<MicrosoftDynamicsCRMmailboxtrackingfolder> ssgCsrspartyMailboxTrackingFolders = default(IList<MicrosoftDynamicsCRMmailboxtrackingfolder>), IList<MicrosoftDynamicsCRMprocesssession> ssgCsrspartyProcessSession = default(IList<MicrosoftDynamicsCRMprocesssession>), IList<MicrosoftDynamicsCRMbulkdeletefailure> ssgCsrspartyBulkDeleteFailures = default(IList<MicrosoftDynamicsCRMbulkdeletefailure>), IList<MicrosoftDynamicsCRMprincipalobjectattributeaccess> ssgCsrspartyPrincipalObjectAttributeAccesses = default(IList<MicrosoftDynamicsCRMprincipalobjectattributeaccess>), IList<MicrosoftDynamicsCRMssgCsrschild> ssgCsrspartySsgCsrschildMother = default(IList<MicrosoftDynamicsCRMssgCsrschild>), IList<MicrosoftDynamicsCRMssgCsrschild> ssgCsrspartySsgCsrschildChildsFather = default(IList<MicrosoftDynamicsCRMssgCsrschild>), IList<MicrosoftDynamicsCRMssgCsrsarchivedpartycontact> ssgCsrspartySsgCsrsarchivedpartycontactPartyName = default(IList<MicrosoftDynamicsCRMssgCsrsarchivedpartycontact>), IList<MicrosoftDynamicsCRMssgCsrsfile> ssgCsrspartySsgCsrsfileRecipient = default(IList<MicrosoftDynamicsCRMssgCsrsfile>), IList<MicrosoftDynamicsCRMssgCsrsfile> ssgCsrspartySsgCsrsfilePayor = default(IList<MicrosoftDynamicsCRMssgCsrsfile>), IList<MicrosoftDynamicsCRMssgCsrsrecalculation> ssgCsrspartySsgCsrsrecalculationRecipient = default(IList<MicrosoftDynamicsCRMssgCsrsrecalculation>), IList<MicrosoftDynamicsCRMssgCsrsrecalculation> ssgCsrspartySsgCsrsrecalculationPayor = default(IList<MicrosoftDynamicsCRMssgCsrsrecalculation>), IList<MicrosoftDynamicsCRMsharepointdocumentlocation> ssgCsrspartySharePointDocumentLocations = default(IList<MicrosoftDynamicsCRMsharepointdocumentlocation>), IList<MicrosoftDynamicsCRMssgCsrsportalaudithistory> ssgSsgCsrspartySsgCsrsportalaudithistoryParty = default(IList<MicrosoftDynamicsCRMssgCsrsportalaudithistory>), IList<MicrosoftDynamicsCRMssgCsrsfeedback> ssgSsgCsrspartySsgCsrsfeedbackCSRSParty = default(IList<MicrosoftDynamicsCRMssgCsrsfeedback>), MicrosoftDynamicsCRMssgCsrsparty ssgDuplicatedParty = default(MicrosoftDynamicsCRMssgCsrsparty), IList<MicrosoftDynamicsCRMssgCsrsparty> ssgCsrspartySsgCsrspartyMasterPartyRecord = default(IList<MicrosoftDynamicsCRMssgCsrsparty>), MicrosoftDynamicsCRMssgCsrsparty ssgLatestDuplicateParty = default(MicrosoftDynamicsCRMssgCsrsparty), IList<MicrosoftDynamicsCRMssgCsrsparty> ssgCsrspartySsgCsrspartyLatestDuplicateParty = default(IList<MicrosoftDynamicsCRMssgCsrsparty>), IList<MicrosoftDynamicsCRMssgCsrscommunicationmessage> ssgCsrspartySsgCsrscommunicationmessageToParty = default(IList<MicrosoftDynamicsCRMssgCsrscommunicationmessage>), IList<MicrosoftDynamicsCRMssgCsrscommunicationmessage> ssgCsrspartySsgCsrscommunicationmessageFromParty = default(IList<MicrosoftDynamicsCRMssgCsrscommunicationmessage>), IList<MicrosoftDynamicsCRMactivityparty> ssgCsrspartyActivityParties = default(IList<MicrosoftDynamicsCRMactivityparty>))
+        public MicrosoftDynamicsCRMssgCsrsparty(int? statecode = default(int?), int? ssgStagingfilenumber = default(int?), string _ssgDuplicatedpartyValue = default(string), System.DateTimeOffset? createdon = default(System.DateTimeOffset?), int? ssgIncomeassistance = default(int?), int? ssgReferral = default(int?), string ssgCellphone = default(string), string ssgFirstname = default(string), int? timezoneruleversionnumber = default(int?), System.DateTimeOffset? modifiedon = default(System.DateTimeOffset?), int? importsequencenumber = default(int?), string ssgCsrspartyid = default(string), string _createdonbehalfbyValue = default(string), int? ssgPartyalreadyexists = default(int?), string _modifiedbyValue = default(string), int? utcconversiontimezonecode = default(int?), System.DateTimeOffset? ssgDateofbirth = default(System.DateTimeOffset?), System.DateTimeOffset? overriddencreatedon = default(System.DateTimeOffset?), string versionnumber = default(string), string ssgLastname = default(string), string ssgReferencenumber = default(string), int? ssgPartygender = default(int?), string ssgBceidUserid = default(string), int? ssgPreferredcontactmethod = default(int?), string ssgCity = default(string), string ssgWorkphone = default(string), int? ssgProvinceterritory = default(int?), string _owningbusinessunitValue = default(string), string _owningteamValue = default(string), string ssgMiddlename = default(string), string ssgFullname = default(string), int? ssgGender = default(int?), string ssgHomephone = default(string), string _owninguserValue = default(string), int? statuscode = default(int?), string _modifiedonbehalfbyValue = default(string), string _ssgLatestduplicatepartyValue = default(string), string ssgStreet2 = default(string), string _createdbyValue = default(string), int? ssgStagingid = default(int?), string _owneridValue = default(string), string ssgAreapostalcode = default(string), string ssgEmail = default(string), System.DateTimeOffset? ssgBceidLastUpdate = default(System.DateTimeOffset?), string ssgPreferredname = default(string), bool? ssgPortalaccess = default(bool?), int? ssgIdentity = default(int?), string ssgBceidDisplayname = default(string), string ssgIdentityotherdetails = default(string), string ssgStreet1 = default(string), string ssgBceidGuid = default(string), MicrosoftDynamicsCRMsystemuser createdby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser createdonbehalfby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser modifiedby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser modifiedonbehalfby = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMsystemuser owninguser = default(MicrosoftDynamicsCRMsystemuser), MicrosoftDynamicsCRMteam owningteam = default(MicrosoftDynamicsCRMteam), MicrosoftDynamicsCRMprincipal ownerid = default(MicrosoftDynamicsCRMprincipal), MicrosoftDynamicsCRMbusinessunit owningbusinessunit = default(MicrosoftDynamicsCRMbusinessunit), IList<MicrosoftDynamicsCRMsyncerror> ssgCsrspartySyncErrors = default(IList<MicrosoftDynamicsCRMsyncerror>), IList<MicrosoftDynamicsCRMduplicaterecord> ssgCsrspartyDuplicateMatchingRecord = default(IList<MicrosoftDynamicsCRMduplicaterecord>), IList<MicrosoftDynamicsCRMduplicaterecord> ssgCsrspartyDuplicateBaseRecord = default(IList<MicrosoftDynamicsCRMduplicaterecord>), IList<MicrosoftDynamicsCRMasyncoperation> ssgCsrspartyAsyncOperations = default(IList<MicrosoftDynamicsCRMasyncoperation>), IList<MicrosoftDynamicsCRMmailboxtrackingfolder> ssgCsrspartyMailboxTrackingFolders = default(IList<MicrosoftDynamicsCRMmailboxtrackingfolder>), IList<MicrosoftDynamicsCRMprocesssession> ssgCsrspartyProcessSession = default(IList<MicrosoftDynamicsCRMprocesssession>), IList<MicrosoftDynamicsCRMbulkdeletefailure> ssgCsrspartyBulkDeleteFailures = default(IList<MicrosoftDynamicsCRMbulkdeletefailure>), IList<MicrosoftDynamicsCRMprincipalobjectattributeaccess> ssgCsrspartyPrincipalObjectAttributeAccesses = default(IList<MicrosoftDynamicsCRMprincipalobjectattributeaccess>), IList<MicrosoftDynamicsCRMssgCsrschild> ssgCsrspartySsgCsrschildMother = default(IList<MicrosoftDynamicsCRMssgCsrschild>), IList<MicrosoftDynamicsCRMssgCsrschild> ssgCsrspartySsgCsrschildChildsFather = default(IList<MicrosoftDynamicsCRMssgCsrschild>), IList<MicrosoftDynamicsCRMssgCsrsarchivedpartycontact> ssgCsrspartySsgCsrsarchivedpartycontactPartyName = default(IList<MicrosoftDynamicsCRMssgCsrsarchivedpartycontact>), IList<MicrosoftDynamicsCRMssgCsrsfile> ssgCsrspartySsgCsrsfileRecipient = default(IList<MicrosoftDynamicsCRMssgCsrsfile>), IList<MicrosoftDynamicsCRMssgCsrsfile> ssgCsrspartySsgCsrsfilePayor = default(IList<MicrosoftDynamicsCRMssgCsrsfile>), IList<MicrosoftDynamicsCRMssgCsrsrecalculation> ssgCsrspartySsgCsrsrecalculationRecipient = default(IList<MicrosoftDynamicsCRMssgCsrsrecalculation>), IList<MicrosoftDynamicsCRMssgCsrsrecalculation> ssgCsrspartySsgCsrsrecalculationPayor = default(IList<MicrosoftDynamicsCRMssgCsrsrecalculation>), IList<MicrosoftDynamicsCRMsharepointdocumentlocation> ssgCsrspartySharePointDocumentLocations = default(IList<MicrosoftDynamicsCRMsharepointdocumentlocation>), IList<MicrosoftDynamicsCRMssgCsrsportalaudithistory> ssgSsgCsrspartySsgCsrsportalaudithistoryParty = default(IList<MicrosoftDynamicsCRMssgCsrsportalaudithistory>), IList<MicrosoftDynamicsCRMssgCsrsfeedback> ssgSsgCsrspartySsgCsrsfeedbackCSRSParty = default(IList<MicrosoftDynamicsCRMssgCsrsfeedback>), MicrosoftDynamicsCRMssgCsrsparty ssgDuplicatedParty = default(MicrosoftDynamicsCRMssgCsrsparty), IList<MicrosoftDynamicsCRMssgCsrsparty> ssgCsrspartySsgCsrspartyMasterPartyRecord = default(IList<MicrosoftDynamicsCRMssgCsrsparty>), MicrosoftDynamicsCRMssgCsrsparty ssgLatestDuplicateParty = default(MicrosoftDynamicsCRMssgCsrsparty), IList<MicrosoftDynamicsCRMssgCsrsparty> ssgCsrspartySsgCsrspartyLatestDuplicateParty = default(IList<MicrosoftDynamicsCRMssgCsrsparty>), IList<MicrosoftDynamicsCRMssgCsrscommunicationmessage> ssgCsrspartySsgCsrscommunicationmessageToParty = default(IList<MicrosoftDynamicsCRMssgCsrscommunicationmessage>), IList<MicrosoftDynamicsCRMssgCsrscommunicationmessage> ssgCsrspartySsgCsrscommunicationmessageFromParty = default(IList<MicrosoftDynamicsCRMssgCsrscommunicationmessage>), IList<MicrosoftDynamicsCRMactivityparty> ssgCsrspartyActivityParties = default(IList<MicrosoftDynamicsCRMactivityparty>))
         {
             Statecode = statecode;
             SsgStagingfilenumber = ssgStagingfilenumber;
@@ -206,7 +206,7 @@ namespace Csrs.Interfaces.Dynamics.Models
         /// <summary>
         /// </summary>
         [JsonProperty(PropertyName = "ssg_dateofbirth")]
-        public string SsgDateofbirth { get; set; }
+        public System.DateTimeOffset? SsgDateofbirth { get; set; }
 
         /// <summary>
         /// </summary>
@@ -216,7 +216,7 @@ namespace Csrs.Interfaces.Dynamics.Models
         /// <summary>
         /// </summary>
         [JsonProperty(PropertyName = "versionnumber")]
-        public int? Versionnumber { get; set; }
+        public string Versionnumber { get; set; }
 
         /// <summary>
         /// </summary>
