+++ src/backend/Csrs.Test/Interfaces/Dynamics/Models/ModelTestBase.cs
@@ -1,0 +1,33 @@
+﻿using Microsoft.Rest.Serialization;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Csrs.Test.Interfaces.Dynamics.Models
+{
+    public abstract class ModelTestBase
+    {
+        protected JsonSerializerSettings JsonSerializerSettings;
+
+        protected ModelTestBase()
+        {
+            JsonSerializerSettings = new JsonSerializerSettings
+            {
+                Formatting = Formatting.None,
+                DateFormatHandling = DateFormatHandling.IsoDateFormat,
+                DateTimeZoneHandling = DateTimeZoneHandling.Utc,
+                NullValueHandling = NullValueHandling.Ignore,
+                ReferenceLoopHandling = ReferenceLoopHandling.Serialize,
+                ContractResolver = new ReadOnlyJsonContractResolver(),
+                Converters = new List<JsonConverter>
+                    {
+                        new Iso8601TimeSpanConverter()
+                    }
+            };
+
+        }
+    }
+}
